requirements.txt
requests
bs4
pandas
pyarrow
minio
python-dotenv